+++ easycv/datasets/download_data/download_voc.py
@@ -2,7 +2,7 @@
 import os
 
 from easycv.utils.constant import CACHE_DIR
-from .commont import check_path_exists, download, extract
+from .commont import check_path_exists, download
 
 cfg = dict(
     voc2007=
@@ -10,6 +10,15 @@ cfg = dict(
     voc2012=
     'http://host.robots.ox.ac.uk/pascal/VOC/voc2012/VOCtrainval_11-May-2012.tar'
 )
+
+
+# Check whether the data exists
+def check_data_exists(name, target_dir, split):
+    root_path = os.path.join(target_dir, 'VOCdevkit', name.upper())
+    if os.path.exists(root_path):
+        return check_path_exists({'path': process_voc(root_path, split=split)})
+    else:
+        return False
 
 
 # return abs path
@@ -20,16 +29,23 @@ def process_voc(path, split='train'):
         return os.path.join(path, 'ImageSets/Main/val.txt')
 
 
+# xtract the data
+def extract(name, file, target_dir=CACHE_DIR):
+    cmd = f'tar -xvf {file} -C {target_dir}'
+    print('begin Unpack.....................')
+    os.system(cmd)
+    print('Unpack is finished.....................')
+
+
 def download_voc(name, split='train', target_dir=CACHE_DIR):
 
-    root_path = os.path.join(target_dir, 'VOCdevkit', name.upper())
-    # 查看此根目录路径是否存在
-    if os.path.exists(os.path.join(root_path)):
-        return check_path_exists({'path': process_voc(root_path, split=split)})
-    # 开始下载
+    if check_data_exists(name, target_dir, split):
+        return check_data_exists(name, target_dir, split)
+
+    # Start the download
     file_path = download(cfg.get(name), target_dir=target_dir)
 
-    # 开始解压
+    # began to unpack
     extract(name, file_path, target_dir=target_dir)
 
-    return check_path_exists({'path': process_voc(root_path, split=split)})
+    return check_data_exists(name, target_dir, split)
